Spawn 0.1 futures when using them as 0.3 futures
This eliminates the need for separate configuration of a global task
when using 0.1 futures on an 0.3 executor.
