+++ futures-util/Cargo.toml
@@ -32,7 +32,7 @@ futures-io-preview = { path = "../futures-io", version = "0.3.0-alpha.7", defaul
 futures-sink-preview = { path = "../futures-sink", version = "0.3.0-alpha.7", default-features = false}
 either = { version = "1.4", default-features = false }
 slab = { version = "0.4", optional = true }
-futures_01 = { version = "0.1", optional = true, package = "futures" }
+futures_01 = { version = "0.1.25", optional = true, package = "futures" }
 tokio-executor = { version = "0.1.2", optional = true }
 tokio-io = { version = "0.1.9", optional = true }
 pin-utils = "0.1.0-alpha.3"
